refactor(types): ExecutorRuntime* 迁到 executor.ts,消除 judge↔report 类型循环
之前 types/judge.ts 用 inline \`import('./report.js').ExecutorRuntimeFingerprint\`
引用,types/report.ts 又顶层 import judge.ts 的 JudgeConfig / JudgeRuntimeEntry,
形成类型层 judge.ts ↔ report.ts 双向循环。types-only 不影响编译/运行,但语义
不对位 — ExecutorRuntimeFingerprint 概念上属于 executor 而非 report。

迁移 8 个相关类型(ExecutorRuntimeKind / ExecutorSystemPromptMode /
ExecutorCostMode / ExecutorTraceMode / ExecutorSkillIsolationMode /
ExecutorRuntimeCapabilities / ExecutorRuntimePackage / ExecutorRuntimeBinary /
ExecutorRuntimeFingerprint)从 types/report.ts 到 types/executor.ts。

- judge.ts: inline 动态 import 改成 top-level \`import type from './executor.js'\`
- report.ts: 顶层 import 加 ExecutorRuntimeFingerprint(原来通过本文件 export 隐式拿到,
  现在显式 import)
- 消费方走 types/index.ts star re-export,零下游改动

依赖图:executor.ts 不依赖 judge.ts / report.ts,变成单向 source。

yarn ci 1031/1031 pass。
